update migrations to use foreignId and constrained methods

+++ database/migrations/2020_08_01_093517_create_accounts_table.php
@@ -18,7 +18,7 @@ class CreateAccountsTable extends Migration
             $table->id();
             $table->string('name');
             $table->boolean('is_active')->default(true);
-            $table->unsignedBigInteger('balance')->default(0);
+            $table->BigInteger('balance')->default(0);
             $table->timestamps();
         });
